+++ otserv/branches/revbattlesys/talkaction.cpp
@@ -173,7 +173,7 @@ long TalkAction::executeSay(Creature* creature, const std::string& words, const 
 	lua_pushstring(L, words.c_str());
 	lua_pushstring(L, param.c_str());
 	
-	long ret;
+	int32_t ret;
 	if(m_scriptInterface->callFunction(3, ret) == false){
 		ret = 0;
 	}
